// src/api/model/itemDto.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemCategoryDto } from './itemCategoryDto';
import { ProviderDto } from './providerDto';


export interface ItemDto { 
    cost?: number;
    id?: number;
    itemCategoryDto?: ItemCategoryDto;
    minQuantity?: number;
    name?: string;
    providerDto?: ProviderDto;
    quantity?: number;
    shelfLife?: number;
    status?: string;
    unit?: string;
}