// src/api/model/productItemDto.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemDto } from './itemDto';
import { ProductDto } from './productDto';


export interface ProductItemDto { 
    id?: number;
    itemDto?: ItemDto;
    productDto?: ProductDto;
    qtde?: number;
}